refactor: modularize predefined valueUnitPairs; add minimum value constraint
